spatio_temporal_nowcasting_tf20/requirements.txt
numpy
pandas
tensorflow-gpu==2.0
matplotlib
opencv-python
lz4
tqdm
h5netcdf
graphviz
pydot
nptyping
wandb
pysolar